Export grn_obj_get_value_()
refs #1496

Note: grn_obj_get_value_() is not recommended to use.
This function is still declared at db.h in this purpose.

for groonga developper:
The groonga.h is installed, but db.h is not.
The symbol (grn_obj_get_value_) is exported.
This constrain a few person who know how to link with can use it.

for groonga user:
This function is not declared at groonga.h,
the advanced user who willing to follow API changes only use it.