gqtp: use network byte order for grn_rc in GQTP header
It breaks backward compatibility. It means that you can not connect to
groonga >= 2.0.9 server with groonga <= 2.0.8 clients.

TODO: We chose whether is this change reasonable or not. Using network
byte order is a good manner for network protocols. But backward
incompatibility is not a good manner for applications. If this change is
not reasonable, we should revert it.