Fix a bug that deleting record can be found
If a record has fixed size column (e.g. Int32) and it is indexed, the
bug is occurred. If a record is deleted, the deleted record can be
found index search with 0 value. 0 value for Int32 is 0.

This fixes mroonga's bug found at
c2b8afce57c89286f8118ebcc5e6260c902c2488.