Use separated array for command error location
CAUTION: This change is backward incompatible change but it will not
break any existing program. Because it is not documented (Sorry. I'm
writing the document.) and it is rare case. No program that parses
output of the case will exist.

Think about the following command in /tmp/error.grn:

  nonexistent_command argument

It is uses with the following command line:

  % groonga --file /tmp/error.grn -n /tmp/db

The output is changed by this change. Here are the outputs with/without
this change. (They are pretty-formatted by hand for readability.)

before:

  [
    [
      -22,
      1353490445.90846,
      0.000262260437011719,
      "invalid command name: nonexistent_command",
      [
        [
          "grn_ctx_qe_exec",
          "ctx.c",
          1394
          "/tmp/error.grn",              # DIFFERENT
          1,                             # DIFFERENT
          "nonexistent_command argument" # DIFFERENT
        ],
      ]
    ]
  ]

after:

  [
    [
      -22,
      1353490445.90846,
      0.000262260437011719,
      "invalid command name: nonexistent_command",
      [
        [
          "grn_ctx_qe_exec",
          "ctx.c",
          1394
        ],                               # DIFFERENT
        [                                # DIFFERENT
          "/tmp/error.grn",              # DIFFERENT
          1,                             # DIFFERENT
          "nonexistent_command argument" # DIFFERENT
        ]
      ]
    ]
  ]