Support vector element cast in grn_obj_set_value()
This changes solves the following case:

    table_create Times TABLE_HASH_KEY Time

    table_create Sites TABLE_HASH_KEY ShortText
    column_create Sites modified_times COLUMN_VECTOR Times

    load --table Sites
    [
    {"_key":"http://groonga.org", ["2013-04-29 00:00:00", "2013-05-02 01:46:48"]}
    ]

Without this change, ["2013-04-29 00:00:00", "2013-05-02 01:46:48"]
are broken. Because these strings aren't casted.

Note that the above example isn't right. Because "load" command casts
["2013-04-29 00:00:00", "2013-05-02 01:46:48"] before "load" command
passes them to grn_obj_set_value(). This problem is caused when
groonga is used as a library such as rroonga use case.