Support cascade delete
You can delete referenced records by this change. You got an error for
the operation before this change.

See the committed test about the detail.

Note that an index must be created for this feature. If no index
exists, the record is just deleted and some dangling references are
appeared.