Revert "Add missing lock for table creation"
This reverts commit 1f1a77402e22a75ca74d488694cd584ac8566dc6.

The lock isn't needed because process global IO objects are guarded
by lock at grn_io_register() and grn_io_unregister() in io.c.

The lock is taken for temporary table that is used for search result.
So the lock may cause a performance regression bug. (We should run
a benchmark but we don't yet...)