ja: improve recycling garbage data
This change increases threshold for big data mode and small data mode.
We treated 127 (= 2 ** 7 - 1) or smaller size data as small data
before this change. Now we treated 65535 (= 2 ** 16 - 1) or smaller
data as small data.

Small data are managed separately for each size range. For example,
[2**0, 2**1) [2**1, 2**2), [2**2, 2**3) and so on. Each separation
keeps datum as an array. It is easy to recycle data but it is less
memory efficiency. For example, think [2**15, 2**16) case. It uses
2 ** 16 size memory for 2 ** 15 + 1 size data. (2 ** 16) - (2 ** 15 +
1) size memory isn't used. This less memory efficiency has more effect
for range range such as [2**15, 2**16).

Large data are managed as packed list. It is difficult to recycle data
but uses memory effectively.

If no data update use case, this change increases database size. This
is a bad case. If you have many small ShortText or COLUMN_VECTOR
columns, you may face a problem.

If many small data update use case, this change keeps database
size. This is a good case.

Normally, increasing database size will be in the allowable range.