grn_output_map_open: change nelements mean
It is incompatible change but some of codes in Groonga don't use the
current behavior correctly. So this incompatible change isn't a
problem.

Before: nelements is sum of the number of keys and the number of values.

After: nelements is the number of elements (= the number of keys or
the number of values).