column_remove: add "dependent" parameter
If "dependent=no" (not default), column_remove reports error for a
column that has one or more indexes.

This change also changes grn_obj_remove(ctx, column) behavior. It
reports error for a column that has one ore more indexes. It's backward
incompatible change.

We need to use grn_obj_remove_dependent(ctx, column) for the previous
behavior.

This is not acceptable. So we need to consider about it carefully.