ii: reduce existing buffer search time
If cursor for searching existing buffer doesn't find any existing buffer
and has many records (= the worst case), many times are spent. It means
that index construction time is slow for the case.

This change limits the number of records for searching existing buffer.

The following case is the worst case:

    table_create a TABLE_NO_KEY
    column_create a x COLUMN_SCALAR ShortText
    table_create b TABLE_PAT_KEY ShortText
    column_create b index COLUMN_INDEX a x

    load --table a
    [
    {"x": "00000000"},
    {"x": "00000001"},
    ...
    {"x": "00486036"},
    {"x": "00000000"},
    {"x": "00000001"},
    ...
    {"x": "00413962"}
    ]