io: don't keep created empty file on error
In the previous versions, there is a case that
empty file keeps remain on error.

Here is the senario to reproduce:

  1. creating new file by grn_fileinfo_open succeeds
  2. mapping file by DO_MAP() is failed

In such a case, it causes an another error such as
"already file exists" because of the file which
isn't under control. so these file should be removed during
cleanup process.