Support falling back query parse
New flag of query_flags in select:

  * QUERY_NO_SYNTAX_ERROR

New API:

  * GRN_EXPR_QUERY_NO_SYNTAX_ERROR

TODO:

  * Document

If QUERY_NO_SYNTAX_ERROR flag is specified as one of query_flags in
select, query never cause syntax error. Syntax error token in query is
parsed as a keyword. For example, "A +" is parsed as "A \+" ("A" and
"+").

This is useful when application uses user input directory and doesn't
want to show syntax error to user and in log.