ii: fix an index break bug
If a chunk is empty by merging, the start record ID gap must be merged
into the next chunk. If we don't merge the start record ID gap, we
can't determine the correct start record ID by chunk info. We need to
read the correct start record ID from the real data. It's not
acceptable for performance.