logical_range_filter: fix exception about closing same object twice
This commit fixes the following error:

  #|e| ArgumentError: already closed object

This exception occurs when max_n_unmatched_records estimation
is not accurate. In such a case, execute_filter(nil) is executed
as a fallback without specifying range index, as a result, no
method exception is raised and closes same object twice in
ensure block.