ii near_phrase: add support for the last phrase by "$"
This is a backward incompatible feature. If "$" is used in near phrase
with Groonga < 10.0.6, the search result is changed.

"$" is chosen because we already use "$" for the last mark such as
"column:$value" in query syntax.

It's only available in near phrase mode. It doesn't work in near
mode. Because introducing "$" breaks compatibility.

Near phrase mode is a recent feature. It will not be used by many
users. Effects of breaking compatibility will be smaller than near
mode. And near phrase mode has escaping syntax. We can escape "$" by
"\$". "\$" returns the same result with old and new Groonga.