TokenDocumentVectorBM25: add a new tokenizer
See the added tests to know how to use.

It generates document vector by BM25:

    IDF = log(((n_documents - DF + 0.5) / (DF + 0.5)) + 1)
    BM25 = IDF * ((TF * (k1 + 1)) /
                  (TF + k1 * (1 - b + b * (DL / AVERAGE_DL))))

The default k1 is 2.0.

The default b is 0.75.

DF and AVERAGE_DL are computed only when the lexicon is empty. If you
change documents, you need to truncate the lexicon and reindex the
indexed column to recompute DF and AVERAGE_DL.