Add support for ordered near phrase search
Script syntax: column *ONP "..."

Query syntax: *ONP"..."

It's similar to near phrase search but all phrases must be ordered
like in query. For example, "a b c" is matched to "a X b Y c" but
isn't matched to "b X a Y c" because "b" is not followed by "a".