grn_table_truncate: don't run reference objects check for temporary table
It must not be referred from other persistent table or column.

It may be referred from other temporary table or column. But it's not
supported yet.

In static index construction, we use temporary lexicon and
grn_table_truncate() for it. It opens all persistent tables and
columns but they never be referred from the temporary lexicon. This
change avoids the needless opens.