query syntax: add support for customizing adjust weight
We need to specify "<" or ">" to all keywords to adjust scores without
this improvement. Because the default adjust weight (6 or 4) is larger
than the default score (1). For example, "A"'s weight is 1 and "B"'s
weight is 4 in "A <B". Decremented "B"'s weight (4) is larger than not
decremented "A"'s weight (1). We need to specify ">A <B" to use
smaller weight than "A" for "B".

We just need to specify "<${WEIGHT}" or ">${WEIGHT}" only to target
keywords with this improvement. For example, "A"'s weight is 1 and
"B"'s weight is 0.9 in "A <0.1B".

There are some incompatibility but it'll be acceptable. We hope that
nobody doesn't use incompatible forms: ">${WEIGHT}...",
"<${WEIGHT}..." and "~${WEIGHT}..."

Before:

  * INCREMENT: 'A >B': "A" AND "B"
    * A's weight: 1
    * B's weight: 6

  * INCREMENT: 'A >0.1B': "A" AND "0.1B": INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * INCREMENT: 'A >"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * DECREMENT: 'A <B': "A" AND "B"
    * A's weight: 1
    * B's weight: 4

  * DECREMENT: 'A <0.1B': "A" AND "0.1B": INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 4

  * NEGATIVE: 'A ~B': "A" ("B" may be contained)
    * A's weight: 1
    * B's weight: -4

  * NEGATIVE: 'A ~0.1B': "A" ("0.1B" may be contained): INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: -4

After:

  * INCREMENT: 'A >0.1B': "A" AND "B": INCOMPATIBLE
    * A's weight: 1
    * B's weight: 1.1

  * INCREMENT: 'A >"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * DECREMENT: 'A <0.1B': "A" AND "B": INCOMPATIBLE
    * A's weight: 1
    * B's weight: 0.9

  * DECREMENT: 'A <"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 4

  * NEGATIVE: 'A ~0.1B': "A" ("B" may be contained): INCOMPATIBLE
    * A's weight: 1
    * B's weight: -0.1

  * NEGATIVE: 'A ~"0.1B"': "A" ("0.1B" may be contained): COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: -4