selector: add "tags" and "tags_column" options
They are common selector options like "score_column".

Records that are matched by a selector with "tags" and "tags_column"
options have the specified "tags" in "tags_column".

For example, the following selector adds a "al-match" tag to "tags"
column of all matched records:

    query("name", "al", {"tags": ["al-match"], "tags_column": "tags"})

If "tags_column" isn't empty, the specified "tags" are appended to the
existing "tags_column".

For example, the former selector adds a "al-match" tag to "tags"
column of all records that have "al" and then the latter selector
appends a "bo-match" tag to "tags" column of all records that have
"bo":

    query("name", "al", {"tags": ["al-match"], "tags_column": "tags"}) ||
    query("name", "bo", {"tags": ["bo-match"], "tags_column": "tags"})

If "tags" is omitted, query itself is used as a tag.

For example, the following selector adds a "al" tag to "tags" column
of all records that have "al" and adds "bo" tag to "tags" column of
all records that have "bo":

    query("name", "al OR bo", {"tags_column": "tags"})

See the added tests for details.