grn_table_selector: add support for weight factor
Weight factor is a factor that is used when logical operator is
processed. The default weight factor is 1.0. It doesn't change weight
of each condition. If weight factor is 2.0, all weights are doubled.

Weight factor is associated with logical
operators (GRN_OP_AND/OR/AND_NOT/ADJUST) but there is no syntax that
uses weight factor. The following commits add it.

New APIs:

  * grn_selector_data_get_weight_factor()
  * grn_table_selector_get_weight_factor()
  * grn_table_selector_set_weight_factor()