NormalizerTable: add (#1214)
See the added tests for how to use.

It uses Groonga table to normalize text. Users can implement custom
normalization by managing a Groonga table.

Note that users need to rebuild the index that uses NormalizerTable
when they change substitution table for the NormalizerTable. Because
normalization rules are changed by change of the substitution table.

Co-authored-by: Takashi Hashida <ellipsoid1989@gmail.com>