grn_ctx_at: add missing lock clear
If some threads open the same object around the same time,
threads except for a thread that executes the opening object
at first are waiting for opening the target object.

At this time, threads that are waited for an opening object
take locks, but these locks are not released.

Therefore, these locks remain until Groonga's process is
restarted in the above case.