snip: add support for regular expression to find delimiter (#1269)
Use case: Use only text in one sentence as snippet by finding "." (for
English) or U+3002 IDEOGRAPHIC FULL STOP (for Japanese).

Note that found delimiters aren't included in snippet. For example,
"." isn't included in snippet with /\./ regular expression. This is
useful for the use case.

This also adds "delimiter_regexp" option to snippet() to use this
feature.