Add a new flag for deciding the type of removed character
I reduce the number of bits for expressing GRN_CHAR_TYPE.
Because I add a new flag.

However, I think that it is enough to ensure 6bit.
Because currently, the kind of GRN_CHAR_TYPE is 9.
We can express 32 pattern GRN_CHAR_TYPE when we can use 6bit for
expressing it.