select: add drilldown_if_max_n_records and drilldowns[].if_max_n_records
If drilldown target table has the number of record larger than or
equal to the specified value, empty drilldown result is returned. It
means that drilldown isn't executed for a large result. Users can omit
drilldown for a large result because it may be expensive.

If negative value is specified, all drilldowns are executed. The
default value is -1. It means that all drilldowns are executed by
default.