ii: add a condition for stopping making the result set on the way when fail to add the result set
Because if Groonga doesn't stop making the result set when failing to
add the result set of the search of regular expressions, Groonga
output many error logs.

Therefore, the command response is slow when the number of hits is
many in this case.

This modification is stopping making the result set on the way when
failing to add the result set.
Thereby, the command response quickly.