hash: reduce the size of the table for setting the result set in advance
Fix [groonga-dev,04992]

Groonga improves performance for setting results set by making the
table for setting the result set in advance.

However, if the number of match records is few, the cost of making the
table is high.
Because the size of the table same as a table of the search targets.

This modifications is made the table for setting the result set in
advance in two steps.

1. Groonga makes the table the size of half a table of the search
   targets before the search execution.

2. If the size of the result sets is over the size of half a table of
   the search targets while the search execution, Groonga expands the
   size of the table for setting the result set in advance.

By doing so, we improve performance if the number of match records is
few.

Reported by Koya Masuda. Thanks!!!