ctx: create temporary objects in top parent grn_ctx
If sub-grn_ctx that is created by grn_ctx_pull_child() creates a
temporary object, the temporary object is created in the top parent
grn_ctx. This may not be a good architecture because grn_ctx is
designed as "one grn_ctx per thread". This breaks it. One grn_ctx (the
top parent grn_ctx) may be used in multiple threads (with lock).

grn_ctx_pull_child() is designed for using sub-grn_ctx in thread
pool. If sub-grn_ctx creates a temporary object such as a filtered
result (grn_hash), dynamic columns (grn_ja and grn_ra) and grouped
result (grn_hash), the temporary object is destroyed in a thread. The
temporary object can't be used out of thread pool. It means that we
can't process large filter/multiple dynamic columns/large group in
parallel with a thread pool. If we create a temporary object in the
top parent grn_ctx, we can process large filter/multiple dynamic
columns/large group in parallel with a thread pool. Because processed
results are owned by the top parent grn_ctx after all processes are
done by a thread pool.