db: add a new condition for stopping result set marge when fail to add the result set
Currently, Groonga doesn't stop making a result set when failing to
add the result set of a OR search.

Therefore, when the number of results of the OR search is many and a
query has many OR conditions, Groonga may response slow with the
"request_cancel" command.

if we execute the "request_cancel" command while Groonga adds the
result set of the OR search, the addition to the result set fails.

Therefore, Groonga output  a error of "failed to add:".
Groonga may output this error a lot because this error occurs every
failing to add the result set.

If Groonga output this error a lot, Groonga's response is very slow.

This modification is stopping making the result set on the way when
failing to add the result set.
Thereby, the command responds quickly.