escalate: add an experimental function
It's similar to the existing match escalation mechanism but this is more
generic.

The existing match escalation mechanism is just for one full text
search by inverted index. But escalate() is for any conditions.

Syntax:

    escalate(THRESHOLD_1, CONDITION_1,
             THRESHOLD_2, CONDITION_2,
             ...,
             THRESHOLD_N, CONDITION_N)

THRESHOLD is a positive number such 0 and 29.

CONDITION is a string that uses script syntax such as "number_column >
29".

THRESHOLD is used for executing the corresponding CONDITION. If the
current result set has less than or equal to THRESHOLD_1 records, the
corresponding CONDITION_1 is executed. If the next result set has less
than or equal to THRESHOLD_2 records, the corresponding CONDITION_2 is
executed. If the next result set has greater than THRESHOLD_3 records,
escalate() is finished.

If all CONDITIONs are executed, "escalate(THRESHOLD_1, CONDITION_1,
..., THRESHOLD_N, CONDITION_N)" is same as "CONDITION_1 || ... ||
CONDITION_N".

escalate() can be worked with logical operators such as "&&" and "&!":

    number_column > 10 && escalate(THRESHOLD_1, CONDITION_1,
                                   ...,
                                   THRESHOLD_N, CONDITION_N)
    number_column > 10 &! escalate(THRESHOLD_1, CONDITION_1,
                                   ...,
                                   THRESHOLD_N, CONDITION_N)

They are same as "number_column > 10 && (CONDITION_1 || ... ||
CONDITION_N)" and "number_column > 10 &! (CONDITION_1 || ... ||
CONDITION_N)". But these behavior may be changed because this behavior
may not be useful.