ii near-phrase-product: fix a not matched bug
It's occurred when multiple 3+ tokens are overlapped in query. For
example, "abc" and "abcd" were an invalid combination.

It's occurred when shorter query is appeared before longer query. For
example, `ONPP1 "(abcd abc) (1 2)"` worked but `ONPP1 "(abc abcd) (1
2)"` didn't work against "abcdx1" .

The search process stopped after shorter one. For example, "abc" and
"1" are compared against "abcdx1". And the interval between "abc" and
"1" is 2 ("dx" and "x1"). 2 is larger than 1 in ONPP1. So the match is
failed. The search process should try "abcd" and "1" but it was not
tried. (The interval between "abcd" and "1" is 1 ("x1") and it's
matched.)