expr: fix a parse bug that "x OR <0.0y" with QUERY_NO_SYNTAX_ERROR
It produces:

    get_value: match_columns
    push: "x"
    match:
    get_value: match_columns
    push: "OR"
    match:
    and:
    get_value: match_columns
    push: "y"
    match:
    and:

not:

    get_value: match_columns
    push: "x"
    match:
    get_value: match_columns
    push: "y"
    or:

It was caused because "OR" and "<0.0" are pended but our pending space
accepts only one pending token. In this case, "OR" is pended and it's
fallback to as a string token instead of logical OR token when "<0.0"
is parsed. Because we want to pend "<0.0" but our pending space is
already used by "OR". We need to process the "OR" in our pending space
to pend "<0.0".

In this change, any pending tokens are accepted instead of falling
back to a string token when "<0.0" (adjust token) is parsed. I hope
that this doesn't cause any syntax error case...

We may want to apply similar change for logical operator cases but I'm
not sure it's syntax error safe. So it's not changed in this change.