select fuzzy_tokenize: add
It's false by default. Because tokenization with TokenMecab will be
failed for typo query. So we will search correct terms from a lexicon
with the given query as-is. Note that typo tolerance search is done
for each keyword not whole query. For example, "Groonga fast" is a
query and "Groonga" and "fast" are keywords (split with spaces).