query_expand: add support for NPP and ONPP
This is an experimental feature.

Before this change, '*NPP"(A) (B) (C)"' is processed as
'*${expand("NPP")}"(${expand("A")}) (${expand("B")}) (${expand("C")})'. (It's
not expected.)

* NPP: Near Phrase Product
* ONPP: Ordered Near Phrase Product

Content for N (Near), NP (Near Phrase) and ONP (Ordered Near Phrase)
isn't expanded at all for now. But we may add support for them. (If we
find a good way to expand...)

Content for NPP and ONPP in expanded like the following:

    *NPP"(A B) (C)" ->
    *NPP"(${expand("A")} ${expand("B")}) (${expand("C")})"

If there is "A" -> ["A1", "A2", "\"A3\""] mapping, expand("A") returns:

    A1 A2 \\"A3\\"

FYI: expand("A") for non NPP/ONPP case:

    ((A1) OR (A2) OR ("A3"))

Note that expand(X) for NPP/ONPP:

  * doesn't use additional "(...)" (Because it's not supported
    in "(...)" for NPP/ONPP)
  * doesn't use OR (Because "(X Y Z)" in NPP/ONPP is "X or Y or Z")
  * escapes ", (, ) and \ (Because expand(X) is always called in
    quote. *NPP"..." and *ONPP need to use "...".)