ja: fix a wrong data created bug
It may be happen with `grn_obj_set_value(numeric_vector_column,
empty_uvector, GRN_OBJ_{APPEND,PREPEND})`.

In general, users can't do this explicitly because the command API
doesn't accept `GRN_OBJ_{APPEND,PREPEND}`. This may be used internally
when a dynamic numeric vector column is created and a temporary result
set is created (OR is used).

For example, the following query may create wrong data:

    select TABLE \
      --match_columns TEXT_COLUMN \
      --query 'A B OR C' \
      --columns[NUMERIC_DYNAMIC_COLUMN].stage result_set \
      --columns[NUMERIC_DYNAMIC_COLUMN].type Float32 \
      --columns[NUMERIC_DYNAMIC_COLUMN].flags COLUMN_VECTOR

If this is happen, `NUMERIC_DYNAMIC_COLUMN` contains many garbage
elements. It also causes too much memory consumption.

Note that this is caused by an uninitialized variable on
stack (`grn_ja_einfo einfo` in `grn_ja_put_raw()` in this case). So
this may or may not be happen.