pat fuzzy: fix a crash bug
It may be happen when:

* query has 2 or more multi-byte characters
* pat has "${ASCII}${ASCII}${MULTIBYTE}*" characters
* WITH_TRANSPOSITION is enabled

For example, a pat that has "aaあ" with query "あああ" pair has this
problem.

To detect transposition, we need to compare

* the current character in query ("x" in the source) and the previous
  character in a pat key ("y" in the source)
* the previous character in query and the current character in pat key

Expression:

    x[i] == y[j - 1] && x[i - 1] == y[j]

In our code, we use character length in bytes instead of offset
because C doesn't support multibyte character. Here is a simplified
problem code:

    memcmp(current_x,
           current_y - current_y_length, /* == previous_y */
           current_x_length) == 0 &&
    memcmp(current_x - current_x_length, /* == previous_x */
           current_y,
           current_y_length) == 0

"- current_?_length" is the problem. We can't use the current length
to compute previous_?. We need to use previous_?_length for it. Here
is a simplified fixed code:

    for (...;
         ...;
         previous_y = current_y, current_y += current_y_length, ...) {
      for (...;
           ...;
           previous_x = current_x, current_x += current_x_length, ...) {
        ...
        memcmp(current_x, previous_y, current_x_length) == 0 &&
        memcmp(previous_x, current_y, current_y_length) == 0
       ...
      }
    }