ci cmake: specify red-arrow version explicitly
I failed to upload deb packages of Apache Arrow 14.0.2 for Ubuntu
20.04 and 22.04. So the official Apache Arrow APT repository doesn't
provide them...

If we don't specify version explicitly, red-arrow 14.0.2 is used and
it requires Apache Arrow GLib 14.0.2. But we can't install it from the
official Apache Arrow APT repository.

It'll be fixed in this month because Apache Arrow 15.0.0 will be
released in this month.