Add applier
It's an optimization like selector. Applier optimizes a dynamic column
creation. We need to initialize values of a dynamic column when we
initialize a dynamic column. In general, each dynamic column value is
initialized by evaluating columns[LABEL].value. An applier generates
values and set them to columns[LABEL] by itself. It will improve
performance by avoiding a common initialization and so on.