normalizers: fix a bug that checks are invalid
If this is happen, highlight may return invalid result because
highlight uses checks.

It's happen when multiple normalizers such as NormalizerTable and
NormalizerNFKC150 are used and REMOVE_BLANK flags is
used (grn_pat_scan(), which is used for highlight, uses it).

grn_string_open_() merges checks by each normalizer. The merge logic
is buggy. If 'NormalizerTable(...), NormalizerNFKC150' is used as
normalizers, checks by NormalizerNFKC150 is overwritten by checks by
NormalizerTable unconditionally. It may produces invalid checks.

The new merge logic is choosing larger check but I'm not sure this is
correct merge logic... We need more cases to consider the correct
merge logic.