table_selector: disable an optimization for large result set and GRN_OP_AND
We have an optimization for GRN_OP_AND and result set that already has
some records. It's faster than the normal operation when result set
doesn't have many records. But it's slower than the normal operation
when result set has many records. We don't use this optimization for
the latter case.

FYI: This optimization was introduced by
69fda634fa875090529a9794680ad8cc655654b6 .