hash: fix a data break bug with about 2^28+ records
grn_hash may rehash with 2^28+ records. If it's happen, grn_hash is
broken or rehash is failed.

Because we need 2^(28+2+1) space to rehash 2^28+ records. We need
2^(28+2) working space to rehash 2^28+ records. We can use 2^(28+1)
working space for 2^28+ but it will not have good computational
efficiency because it will have many conflict. So we use 2^(28+2)
working space. We copy the original 2^28+ records to the new working
space. So we need 2^(28+2+1) space in total.

But we have only 2^30 space in total. So it's happen, grn_hash is
broken or rehash is failed.

We should not rehash again when grn_hash is already rehashed for 2^27+
records.

This also checks table full in grn_hash_add_entry() to support
grn_hash_add() with existing key. If we check table full in
grn_hash_ensure_rehash(), we'll report an error for grn_hash_add()
with existing key too.