TokenNgram: fix a wrong source_first_character_length bug
It's happen when the first character is a blank. We can't detect that
there is any blank characters before the first normalized
character. Because checks[-1] not checks[0] has GRN_STR_BLANK
flag. There is no checks[-1] for the first normalized character.

We need to handle the first normalized character as the special case.

TODO: This logic isn't perfect for the case that has both of leading
and trailing spaces around the first character such as " a ". But
we'll use this logic for now because this will not be a problem for
highlighting.