Add rc check after grn_scan_scan_info_build_simple() at grn_scan_info_build()
Add rc check because the error message changes depending on whether mruby is
enabled or disabled when rc is not `GRN_SUCCESS`.

When mruby is enabled and rc is not `GRN_SUCCESS`, NULL is retuned because rc
is checked.
(The next process will not continue.)

If mruby is disabled and rc is not checked, the next process will be executed
and the error message may be overwritten.