packages debian: prevent attempt made to download or upload any content
We want to prevent attempt made to download or upload any content on CI.

Currently, we build packages for Ubuntu on Lanchpad.
However, we can't download or upload with external hosts on Lanchpad.

If We prevent attempt made to download or upload any content on CI by
this PR, we detect missing dependency pacakges in advance.