packages almalinux: add --platform linux/arm64 when building arm64's image
ref: https://github.com/groonga/groonga/issues/1806

This change solves the following error on CI.
Because Docker build process does not automatically detect the correct platform for the image.
As a result, the platform information defaults to the host platform.
So in this case, platform infromation wasn't the arm64.

To resolve this issue, we need to specify the platform explicitly in the Docker build command.
ref: https://docs.docker.com/build/building/multi-platform/

```

------
 > [internal] load metadata for docker.io/arm64v8/almalinux:8:
------
Dockerfile:2
--------------------
   1 |     ARG FROM=almalinux:8
   2 | >>> FROM ${FROM}
   3 |
   4 |     ARG DEBUG
--------------------
ERROR: failed to solve: arm64v8/almalinux:8: failed to resolve source metadata for docker.io/arm64v8/almalinux:8: no match for platform in manifest: not found
rake aborted!#1 [internal] load build definition from Dockerfile
```

ref: ref: https://github.com/groonga/groonga/actions/runs/9705591558/job/26788067581