grn_obj_set_info_source: use more crash safe save order
Before:
1. Save hooks of sources
2. Save sources of the index column

After:
1. Save sources of the index column
2. Save hooks of sources

If Groonga is crashed between 1. and 2. with the before order, there
are hooks of sources but index column doesn't have sources. In WAL
recovery process, index column will be regenerated by removing the
index column and create a new index column. In the "removing the index
column" process, hooks of sources aren't removed. Because the removing
process finds hooks of sources by sources of the index column. If
hooks of sources are remained, they break something because they are
dangling references.

If we use the after order, dangling hooks of sources aren't remained.