wal recover: introduce removing mark to ensure removing an object
We needs the following steps on removing an object:
1. Remove paths for the object
2. Remove metadata for the object

If Groonga is crashed between 1. and 2., WAL recovery process can't
determine whether the object should be recovered or removed.

This creates "#{DB}.#{id}.removing" path before 1. and removes it
after 2. If there is the path, WAL recovery process removes the object
and the path instead of trying recovering the object.