Add the comment about the precondition
This function assumes that the input utf8_char is a valid UTF-8 character.
It is the caller's responsibility to ensure that utf8_char is valid UTF-8
before calling this function. This is a precondition in terms of contract programming.

Although UTF-8 characters starting with 0xE1 should be three bytes long, this function
is called under the condition (unified_char_length == 2 || unified_char_length == 3),
so it is possible for a corrupted UTF-8 character to be two bytes long.

However, any corrupted UTF-8 characters should be removed by grn_charlen_() before
reaching this function, so it is assumed that the input is valid UTF-8.